Insert safeguard in case of incorrect username
An incorrect username sets @user to nil.
Previously, this would cause a NoMethodError
when calling @user.authenticate. The newly
formatted conditional prevents this method
from being called if @user is nil.

+++ app/controllers/sessions_controller.rb
@@ -4,7 +4,7 @@ class SessionsController < ApplicationController
 
   def create
     @user = User.find_by_username(params[:username])
-    if @user.authenticate(params[:password])
+    if @user && @user.authenticate(params[:password])
       login @user
       redirect_to session.delete(:return_to) || root_path
     else
